# internal/app/app-api/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server API.",
        "title": "APP Ecommerce API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/order/history": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "admin order get history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "admin order get history",
                "operationId": "OrderGetHistoryAdmin",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "OrderGetHistoryReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.OrderGetHistoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderGetHistoryRes"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "auth login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "auth login",
                "operationId": "AuthLogin",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "AuthLoginReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.AuthLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "auth logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "auth logout",
                "operationId": "AuthLogout",
                "responses": {}
            }
        },
        "/api/auth/refresh": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "auth refresh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "auth refresh",
                "operationId": "AuthRefresh",
                "responses": {}
            }
        },
        "/api/auth/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "auth register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "auth register",
                "operationId": "AuthRegister",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "AuthRegisterReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.AuthRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/user-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "auth user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "auth user info",
                "operationId": "AuthUserInfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.AuthUserInfo"
                        }
                    }
                }
            }
        },
        "/api/order/cancel/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "orderCancel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order Cancel",
                "operationId": "OrderCancel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "OrderCancelReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.OrderCancelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderCancelRes"
                        }
                    }
                }
            }
        },
        "/api/order/completed/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "orderCompleted",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order Completed",
                "operationId": "OrderCompleted",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderCompletedRes"
                        }
                    }
                }
            }
        },
        "/api/order/confirm/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "orderConfirm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order confirm",
                "operationId": "OrderConfirm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderConfirmRes"
                        }
                    }
                }
            }
        },
        "/api/order/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "order create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order create",
                "operationId": "OrderCreate",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "OrderCreateReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.OrderCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderCreateRes"
                        }
                    }
                }
            }
        },
        "/api/order/history": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "order get history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order get history",
                "operationId": "OrderGetHistory",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "OrderGetHistoryReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.OrderGetHistoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderGetHistoryRes"
                        }
                    }
                }
            }
        },
        "/api/order/shipping/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "orderShipping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order Shipping",
                "operationId": "OrderShipping",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderShippingRes"
                        }
                    }
                }
            }
        },
        "/api/order/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "orderdelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "order delete",
                "operationId": "OrderDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.OrderDeleteRes"
                        }
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "product get list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "product get list",
                "operationId": "ProductGetList",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "ProductGetListReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.ProductGetListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ProductGetListRes"
                        }
                    }
                }
            }
        },
        "/api/sse": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "SSE Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "SSE"
                ],
                "summary": "SSE Event",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/todo/ping": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "todo ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "todo ping",
                "operationId": "TodoPing",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.AuthLoginReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "data.AuthRegisterReq": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "data.AuthUserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "lastPing": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "data.OrderCancelReq": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "data.OrderCancelRes": {
            "type": "object",
            "properties": {
                "modifiedBy": {
                    "type": "string"
                },
                "modifiedDate": {
                    "type": "string"
                }
            }
        },
        "data.OrderCompletedRes": {
            "type": "object",
            "properties": {
                "modifiedBy": {
                    "type": "string"
                },
                "modifiedDate": {
                    "type": "string"
                }
            }
        },
        "data.OrderConfirmRes": {
            "type": "object",
            "properties": {
                "modifiedBy": {
                    "type": "string"
                },
                "modifiedDate": {
                    "type": "string"
                }
            }
        },
        "data.OrderCreateReq": {
            "type": "object",
            "required": [
                "orderDetails"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "orderDetails": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "$ref": "#/definitions/data.OrderDetail"
                    }
                }
            }
        },
        "data.OrderCreateRes": {
            "type": "object",
            "properties": {
                "createBy": {
                    "type": "string"
                },
                "createCount": {
                    "type": "integer"
                },
                "createdDate": {
                    "type": "string"
                },
                "createdID": {
                    "type": "integer"
                },
                "deleteCount": {
                    "type": "integer"
                },
                "updateCount": {
                    "type": "integer"
                }
            }
        },
        "data.OrderDeleteRes": {
            "type": "object",
            "properties": {
                "deletedBy": {
                    "type": "string"
                },
                "deletedCount": {
                    "type": "integer"
                },
                "deletedDate": {
                    "type": "string"
                }
            }
        },
        "data.OrderDetail": {
            "type": "object",
            "required": [
                "productID",
                "quantity"
            ],
            "properties": {
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "data.OrderDetailResult": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "proudctName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "data.OrderGetHistoryReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 10
                },
                "sortBy": {
                    "type": "object",
                    "properties": {
                        "field": {
                            "type": "string",
                            "example": "modifiedDate"
                        },
                        "mode": {
                            "type": "string",
                            "example": "desc"
                        }
                    }
                },
                "status": {
                    "type": "string",
                    "example": ""
                },
                "userID": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "data.OrderGetHistoryRes": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.OrderGetHistoryResult"
                    }
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalResults": {
                    "type": "integer"
                }
            }
        },
        "data.OrderGetHistoryResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "modifiedDate": {
                    "type": "string"
                },
                "orderDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.OrderDetailResult"
                    }
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "data.OrderShippingRes": {
            "type": "object",
            "properties": {
                "modifiedBy": {
                    "type": "string"
                },
                "modifiedDate": {
                    "type": "string"
                }
            }
        },
        "data.ProductGetListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 10
                },
                "searchText": {
                    "type": "string",
                    "maxLength": 255,
                    "example": ""
                },
                "sortBy": {
                    "type": "object",
                    "properties": {
                        "field": {
                            "type": "string",
                            "example": "id"
                        },
                        "mode": {
                            "type": "string",
                            "example": "asc"
                        }
                    }
                }
            }
        },
        "data.ProductGetListRes": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.ProductGetListResult"
                    }
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalResults": {
                    "type": "integer"
                }
            }
        },
        "data.ProductGetListResult": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}